requirements.txt
amqp==5.0.6
anyjson==0.2.5
arabic-reshaper==2.1.3
asgiref==3.4.1
billiard==3.6.4.0
celery==5.2.1
certifi==2021.10.8
chardet==3.0.4
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==3.2.9
djangorestframework==3.12.4
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
idna==2.8
kombu==5.2.2
Pillow==8.4.0
prompt-toolkit==3.0.22
psycopg2-binary==2.9.2
PyPDF2==1.26.0
python-bidi==0.4.2
python-dotenv==0.19.2
pytz==2021.3
reportlab==3.6.2
requests==2.21.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.24.3
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.13.3
xhtml2pdf==0.2.5
